extract PATH lookup so it can be reused
